chore(lint): work the ruff backlog from 130 to 51, each finding read first
Ruff's safe fixes (unused imports and noqa, chained operators parenthesised
the way Python already binds them, sorted __slots__/__all__), then by hand:

- zip(strict=True) where a loop ASSERTS the pairs agree: the parity corpus's
  ranking keys and delta lists, the golden-replay checkpoints, the evaluator
  tests. A plain zip compares the overlap and passes on a length mismatch,
  which in the exactness oracle is a test that cannot fail (shape 8). Also
  in src where lengths are equal by construction: dice faces, the
  evaluator's 2-tuples, the opponent model's layers (a malformed checkpoint
  now raises instead of truncating).
- dead variables removed; loop variables bound explicitly in closures; files
  opened with a context manager; to_ops walks itertools.pairwise; the early-
  stopping trial loop is all(), which short-circuits at the same draw.
- ty: _hand_attack_value can return None and region_vp returns a float, so
  the annotations say so. ty 113 -> 109 at the default config.

Left, deliberately: the five bonus-region zips in core.py, rollout.py and
policy.py pair `non_bonus or ()` with `bonus`, and strict there could
raise mid-game where today it stops at the shorter; the two SIM103 rewrites
would break the parallel shape of coup_forbidden's checks; RUF005 and
PLW1510 are style. The two ty: ignore directives in events.py look unused
at the default config but are used by tests/test_types.py's gated config --
removing them failed that test, and they are restored.

Full suite: 774 passed, 3 skipped, 2 xfailed, and the one failure was those
two directives, now back to HEAD; test_types passes. The parity corpus
passing is what says no bot decision moved.

+++ src/struggler/bots/greedy.py
@@ -162,8 +162,8 @@ def _expected_coup_gain(
     modifier = coup_roll_modifier_estimate(observation, side, info)
     expected_margin = 3.5 + ops - 2 * info.stability + modifier
     opp_inf = board.influence[country][opponent.value]
-    opp_removed = int(round(max(0.0, min(expected_margin, opp_inf))))
-    leftover = int(round(max(0.0, expected_margin - opp_removed)))
+    opp_removed = round(max(0.0, min(expected_margin, opp_inf)))
+    leftover = round(max(0.0, expected_margin - opp_removed))
 
     before = board_value(weights, board, side)
     board.influence[country][opponent.value] -= opp_removed
@@ -229,12 +229,12 @@ def _score_realignment_target(
 
     before = board_value(weights, board, side)
     if expected_margin > 0:
-        removed = int(round(min(expected_margin, board.influence[country][opponent.value])))
+        removed = round(min(expected_margin, board.influence[country][opponent.value]))
         board.influence[country][opponent.value] -= removed
         after = board_value(weights, board, side)
         board.influence[country][opponent.value] += removed
     elif expected_margin < 0:
-        removed = int(round(min(-expected_margin, board.influence[country][side.value])))
+        removed = round(min(-expected_margin, board.influence[country][side.value]))
         board.influence[country][side.value] -= removed
         after = board_value(weights, board, side)
         board.influence[country][side.value] += removed
